# WIP make symbolication async