--- a/README.md
+++ b/README.md
@@ -52,7 +52,8 @@ use minidump::Minidump;
 use minidump_processor::{http_symbol_supplier, ProcessorOptions, Symbolizer};
 use serde_json::Value;
 
-fn main() -> Result<(), ()> {
+#[tokio::main]
+async fn main() -> Result<(), ()> {
     // Read the minidump
     let dump = Minidump::read_path("../testdata/test.dmp").map_err(|_| ())?;
  
@@ -77,6 +78,7 @@ fn main() -> Result<(), ()> {
     ));
  
     let state = minidump_processor::process_minidump_with_options(&dump, &provider, options)
+        .await
         .map_err(|_| ())?;
 
     // Write the JSON output to an arbitrary writer (here, a Vec).