--- a/minidump-processor/Cargo.toml
+++ b/minidump-processor/Cargo.toml
@@ -20,6 +20,7 @@ breakpad-syms = ["breakpad-symbols"]
 symbolic-syms = []
 
 [dependencies]
+async-trait = "0.1.51"
 breakpad-symbols = { version = "0.9.6", path = "../breakpad-symbols", optional = true }
 clap = "2.34"
 log = "0.4"
@@ -34,3 +35,4 @@ thiserror = "1.0.30"
 test-assembler = "0.1.6"
 synth-minidump = { path = "../synth-minidump" }
 doc-comment = "0.3.3"
+tokio =  { version = "1.12.0", features = ["full"] }