I was curious how bad this was and then suddenly I was "done" lol

This compiles and correctly handles amd64.

TODO: 

* [x] copy-paste changes to other stackwalking backends (and re-enable them)
* [x] make tests work with async (with `#[tokio::test]`?)
* [x] make sure the symbolic-syms shim still compiles
* [x] performance analysis (socc-pair --bench)
* ~~[ ] test `#[tokio::main(flavor = "current_thread")]`~~ (perf analysis came back good enough that I don't care)