@Gankra 2021-12-01
implements #327