@Gankra 2021-12-01
Me: what if I add Send + Sync everywhere. Surely that won't work. There's like, borrows and stuff right?

rustc: seems fine

Me: well, this is the entire purpose for your existence so I'm going to assume this is correct..?

---------

I feel like I'm probably not full leveraging the power of async in fetch_symbol_file and save_contents, but that's probably better left to the sentry folks who actually use async all the time.