@luser 2021-12-01
> Me: what if I add Send + Sync everywhere. Surely that won't work. There's like, borrows and stuff right?

The minidump data is only ever borrowed immutably. :) At one point I pondered what a multithreaded `process_minidump` implementation would look like but I discarded that idea because the actual unwinding takes no time at all. If anything, per the Discord discussion we had today, doing all the unwinding in one pass, fetching all the symbol files needed to symbolicate in parallel, and then symbolicating the stacks in one fell swoop would be the most clever approach. (The only downside being that the current heuristics for stack scanning rely on symbols to determine if a return address is inside a known function.)