@Gankra 2021-12-02
As far as I know, this is a complete implementation. Still not gonna land it anytime this ~month ideally but it's here if anyone wants to build on it.