@Gankra 2021-12-02
cc @nornagon in case this is at all still relevant to you