@nornagon 2021-12-08
Oh cool! Is the conversion to the new symbolizer done too?