@Gankra 2021-12-08
Unfortunately the folks at sentry decided to focus on something else, and I had higher priorities to take care of first. Also in local testing we've still found this implementation is still significantly faster than breakpad, which is nice. 

Now that rust-minidump is ready for production, I checked in with sentry and they cited the lack of async symbolication support as an issue. So this is in some sense "preparing" for the eventual future where they're the backend. But for the short-to-midterm breakpad-symbols is gonna be The Backend.

Note that mozilla is in the middle of transitioning our infrastructure to rust-minidump, so I'm holding off on landing this patch until probably january when everyone's back from the holidays, to minimize risk. In the meantime, this branch is here for you to build on top of if you want (I imagine you still need to add a bit more async to make minidump loading work for you, and to add a bunch of wasm cfgs). If you'd like to touch base on what electron needs from this, I'm happy to discuss further details on whatever your preferred platform is.