@Gankra 2021-12-17
Rebased, but this is effectively reverting the functionality of #359 right now. I need to rework the code to use `Response::chunks` which is the "async" version of Read for reqwest. This is slightly annoying because File i/o and what not does not use `chunks`. Might just copy-paste the code, idk.