@Gankra 2021-12-17
Async streaming parsing now works? Kinda sloppy but seems legit?

```
miniump-stackwalk runtime: 00m:06s:885ms (6885ms)
miniump-stackwalk max memory (rss): 255MB (268423168 bytes)
```