@Gankra 2021-12-17
I think that's a wrap for me in 2021, see y'all in January.