@Gankra 2021-12-17
Ok one last note before I forget:

In this design, there's kind of a split async personality:

* When reading from a file, once you start reading you're in a blocking loop that will run to completion.
* When downloading a file, the process is fully async, and will `await` any time we need more data from the network.

This is intuitively the right balance since async file I/O is a Lie, but I don't know enough about what people "want" from async to know for sure. That said it could all be moot in a world of Symbolic, because all this code gets banished to hell.

Although I could imagine this "version" being friendlier for compile-to-wasm? So maybe we do end up keeping both paths around. *shrug*