@Gankra 2022-01-12
I added a --mock-server mode to `socc-pair`, so I can now bench the server path without actually hitting the network:

```
socc-pair
--api-token=<snip>
--run-local=C:\Users\ninte\dev\rust-minidump\minidump-stackwalk 
--crash-id=9520523c-45ff-4d6c-a186-4fec80211216 
--output-file=output.diff 
--mock-server 
--bench=30
```

# Summary of Results

## This Change


* average runtime: 1602ms -> 1643ms (1.02x runtime, probably within margin of error)
* average max-memory: 253MB -> 254MB (no significant change)

## Remeasurement of streaming parsing change


* average runtime: 1864ms -> 1602ms (0.86x runtime, nice little win but easily dwarfed by real download times)
* average max-memory: 383MB -> 253MB (0.66x memory usage!)


# Individual Results:

## async (this PR) - 30 iterations, cold cache (local static file server):

```
benchmark results (ms):
  1630, 1642, 1604, 1645, 1645, 1655, 1649, 1626, 1684, 1641, 1633, 1652, 1628, 1624, 1622, 1660, 1640, 1663, 1599, 1717, 1648, 1674, 1671, 1627, 1626, 1666, 1624, 1624, 1643, 1642, 
average runtime: 00m:01s:643ms (1643ms)
median runtime: 00m:01s:642ms (1642ms)
min runtime: 00m:01s:599ms (1599ms)
max runtime: 00m:01s:717ms (1717ms)

max memory (rss) results (bytes):
  266653696, 266702848, 266485760, 265920512, 267214848, 266309632, 267427840, 266227712, 266633216, 265838592, 266575872, 266588160, 265940992, 266084352, 266993664, 266498048, 265887744, 266072064, 266149888, 266407936, 266211328, 266932224, 266964992, 267255808, 266276864, 266936320, 266256384, 267132928, 265351168, 267161600, 
average max-memory: 254MB (266503099 bytes)
median max-memory: 254MB (266498048 bytes)
min max-memory: 253MB (265351168 bytes)
max max-memory: 255MB (267427840 bytes)
```

```

## master - 30 iterations, cold cache (local static file server):

  1610, 1612, 1574, 1612, 1600, 1568, 1583, 1594, 1590, 1607, 1625, 1581, 1610, 1799, 1599, 1592, 1606, 1594, 1597, 1601, 1571, 1586, 1590, 1596, 1592, 1589, 1603, 1595, 1577, 1631, 
average runtime: 00m:01s:602ms (1602ms)
median runtime: 00m:01s:596ms (1596ms)
min runtime: 00m:01s:568ms (1568ms)
max runtime: 00m:01s:799ms (1799ms)

max memory (rss) results (bytes):
  265908224, 266510336, 265248768, 265646080, 265555968, 265035776, 266092544, 265515008, 265510912, 265961472, 266174464, 264835072, 265707520, 265990144, 266002432, 266915840, 266076160, 265310208, 265334784, 266268672, 265588736, 265003008, 265302016, 265940992, 265338880, 265875456, 265576448, 265646080, 265846784, 265408512, 
average max-memory: 253MB (265704243 bytes)
median max-memory: 253MB (265646080 bytes)
min max-memory: 252MB (264835072 bytes)
max max-memory: 254MB (266915840 bytes)
```




## before streaming parser (2dc7ecc384f8ed123387cace30157c5288) - 30 iterations, cold cache (local static file server):

```
benchmark results (ms):
  1771, 1897, 1815, 1888, 2049, 2035, 1733, 1932, 1736, 1751, 1890, 1742, 1912, 1790, 1881, 2069, 1911, 1724, 1762, 1855, 1832, 2029, 1915, 1932, 1863, 1872, 1830, 1762, 1854, 1903, 
average runtime: 00m:01s:864ms (1864ms)
median runtime: 00m:01s:872ms (1872ms)
min runtime: 00m:01s:724ms (1724ms)
max runtime: 00m:02s:069ms (2069ms)

max memory (rss) results (bytes):
  402206720, 401584128, 402771968, 400928768, 401539072, 402026496, 401924096, 402530304, 402178048, 402096128, 402280448, 401825792, 401575936, 402894848, 402010112, 401981440, 401850368, 402341888, 401457152, 402329600, 401281024, 401469440, 401362944, 402178048, 402026496, 402587648, 402583552, 402243584, 401514496, 402354176, 
average max-memory: 383MB (401997824 bytes)
median max-memory: 383MB (402026496 bytes)
min max-memory: 382MB (400928768 bytes)
max max-memory: 384MB (402894848 bytes)
```