@Gankra 2022-02-02
@jan-auer I'm inclined to land this so it stops being a looming specter that I keep needing to rebase. Any final objections / arguments against it?